Set version to 0.3.1.20171023.1
The upstream library does not really release any versions. The latest
version they released was 0.3.0 in 2014, but we are using code from
quite a bit newer than that. The older version we added was 0.4.0, but
this does not fit in with our third-party versioning scheme. To fit in
with the third-party versioning scheme, we are adding a fake release for
upstream at 8e4fc4e, setting the version to 0.3.1.20171023 (adding a
patch bump to 0.3.0 and adding a date string for it as well). We then
release our Starry changes by adding a new component to the version,
which gives us 0.3.1.20171023.1.